Role name colored in dialog
